refactor: Move the utils feature lib to the shared workspace lib
