+++ packages/shared/utils/component-data.util.test.ts
