+++ packages/shared/utils/component-data.util.ts
