+++ packages/shared/utils/flatten.util.ts
