+++ src/renderer/src/features/designer/features/data/manager/AvailableComponentsManager.tsx
@@ -1,10 +1,10 @@
 import { ComponentData } from '@kitae/shared/types'
+import { getSlots } from '@kitae/shared/utils'
 import Accordion from '@renderer/components/Accordion'
 import Icon from '@renderer/components/Icon'
 import { defaultComponents } from '@renderer/features/designer/default-components'
 import { componentTypeIconMap } from '@renderer/features/designer/icon-map'
 import { Draggable, draggable, droppable } from '@renderer/features/drag-n-drop'
-import { getSlots } from '@renderer/features/utils/slot.util'
 import { Component, For, JSX, Show, createMemo } from 'solid-js'
 import { twMerge } from 'tailwind-merge'
 import { useDesignerState } from '../../state/designer.state'
