+++ src/renderer/src/features/designer/features/data/manager/ColorsManager.tsx
@@ -1,9 +1,9 @@
 import { Path } from '@kitae/shared/types'
+import { samePath } from '@kitae/shared/utils'
 import Badge from '@renderer/components/Badge'
 import Button from '@renderer/components/Button'
 import AddInput from '@renderer/components/form/AddInput'
 import Icon from '@renderer/components/Icon'
-import { samePath } from '@renderer/features/utils/same-path.util'
 import { Component, createEffect, createMemo, createSignal, For, JSX } from 'solid-js'
 import { makeAddThemeEntryChange } from '../../history/theme.events'
 import { useDesignerState } from '../../state/designer.state'
