+++ src/renderer/src/features/designer/features/properties/forms/ComponentTextProperty.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { ComponentConfig } from '@kitae/shared/types'
+import { walker } from '@kitae/shared/utils'
 import FormField from '@renderer/components/form/FormField'
 import { createForm } from '@renderer/features/form'
-import { walker } from '@renderer/features/utils/walker.util'
 import { debounce } from '@solid-primitives/scheduled'
 import { Component, createEffect, createMemo } from 'solid-js'
 import { makeUpdateTextConfigPropertyChange } from '../../history/property.events'
