+++ src/renderer/src/features/designer/features/properties/forms/ThemeColorValueProperty.tsx
@@ -1,8 +1,8 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
+import { walker } from '@kitae/shared/utils'
 import ColorPicker from '@renderer/components/form/ColorPicker'
 import FormField from '@renderer/components/form/FormField'
 import { createForm } from '@renderer/features/form'
-import { walker } from '@renderer/features/utils/walker.util'
 import { debounce } from '@solid-primitives/scheduled'
 import Color from 'color'
 import { Component, createEffect } from 'solid-js'
